requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile --strip-extras
#
asgiref==3.8.1
    # via opentelemetry-instrumentation-asgi
deprecated==1.2.14
    # via opentelemetry-api
fastcgi==0.0.3
    # via -r requirements.in
fastcore==1.5.46
    # via fastcgi
importlib-metadata==7.1.0
    # via opentelemetry-api
opentelemetry-api==1.25.0
    # via
    #   -r requirements.in
    #   opentelemetry-instrumentation
    #   opentelemetry-instrumentation-asgi
    #   opentelemetry-instrumentation-fastapi
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
opentelemetry-instrumentation==0.46b0
    # via
    #   opentelemetry-instrumentation-asgi
    #   opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-asgi==0.46b0
    # via opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-fastapi==0.46b0
    # via -r requirements.in
opentelemetry-sdk==1.25.0
    # via -r requirements.in
opentelemetry-semantic-conventions==0.46b0
    # via
    #   opentelemetry-instrumentation-asgi
    #   opentelemetry-instrumentation-fastapi
    #   opentelemetry-sdk
opentelemetry-util-http==0.46b0
    # via
    #   opentelemetry-instrumentation-asgi
    #   opentelemetry-instrumentation-fastapi
packaging==24.1
    # via
    #   fastcgi
    #   fastcore
typing-extensions==4.12.2
    # via
    #   asgiref
    #   opentelemetry-sdk
wrapt==1.16.0
    # via
    #   deprecated
    #   opentelemetry-instrumentation
zipp==3.19.2
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
